feat(T1.5): extract frontApp.js from index.js monolith
Move all routes, middleware, helpers, multer/session config and DB init
into server/src/apps/frontApp.js (lift-and-shift, no behaviour change).
index.js is now a 7-line thin entrypoint that requires the app and calls
listen(PORT). Path constants (__dirname, dotenv, EJS views) adjusted for
the new file location.

This unblocks T1.6 (adminApp shell) and T1.7 (dual-port startup).

Closes #9
